Fixed a pair of TODOs.
One was about the format of samples: it has been changed to accept single point
floating point numbers, in native endian.

The other one is about the period as integer number, which is required in
detect.c: an output parameter has been added to estimatePeriod, to get it
without interpolations and/or fixes.